+++ main_test.go
@@ -1,6 +1,7 @@
 package main
 
 import (
+	"flag"
 	"fmt"
 	"io/ioutil"
 	"math"
@@ -30,7 +31,7 @@ diagnostic=false
 	confFile.Sync()
 	confFile.Close()
 	defer os.Remove(confFile.Name())
-	mergedConfig, _ := resolveConfig([]string{"-conf=" + confFile.Name(), "-role=My-Service:default,INVALID#SERVICE", "-verbose", "-diagnostic"})
+	mergedConfig, _ := resolveConfig(&flag.FlagSet{}, []string{"-conf=" + confFile.Name(), "-role=My-Service:default,INVALID#SERVICE", "-verbose", "-diagnostic"})
 
 	t.Logf("      apibase: %v", mergedConfig.Apibase)
 	t.Logf("       apikey: %v", mergedConfig.Apikey)
@@ -57,30 +58,6 @@ diagnostic=false
 	}
 }
 
-func TestParseFlagsPrintVersion(t *testing.T) {
-	config, otherOptions := resolveConfig([]string{"-version"})
-
-	if config.Verbose != false {
-		t.Error("with -version args, variables of config should have default values")
-	}
-
-	if otherOptions.printVersion == false {
-		t.Error("with -version args, printVersion should be true")
-	}
-}
-
-func TestParseFlagsRunOnce(t *testing.T) {
-	config, otherOptions := resolveConfig([]string{"-once"})
-
-	if config.Verbose != false {
-		t.Error("with -version args, variables of config should have default values")
-	}
-
-	if otherOptions.runOnce == false {
-		t.Error("with -once args, RunOnce should be true")
-	}
-}
-
 func TestDetectForce(t *testing.T) {
 	// prepare dummy config
 	confFile, err := ioutil.TempFile("", "mackerel-config-test")
@@ -94,7 +71,7 @@ func TestDetectForce(t *testing.T) {
 	defer os.Remove(confFile.Name())
 
 	argv := []string{"-conf=" + confFile.Name()}
-	conf, force, err := resolveConfigForRetire(argv)
+	conf, force, err := resolveConfigForRetire(&flag.FlagSet{}, argv)
 	if force {
 		t.Errorf("force should be false")
 	}
@@ -103,7 +80,7 @@ func TestDetectForce(t *testing.T) {
 	}
 
 	argv = append(argv, "-force")
-	conf, force, err = resolveConfigForRetire(argv)
+	conf, force, err = resolveConfigForRetire(&flag.FlagSet{}, argv)
 	if !force {
 		t.Errorf("force should be true")
 	}
@@ -136,7 +113,7 @@ func TestResolveConfigForRetire(t *testing.T) {
 		"-role=hoge:fuga",
 	}
 
-	conf, force, err := resolveConfigForRetire(argv)
+	conf, force, err := resolveConfigForRetire(&flag.FlagSet{}, argv)
 	if force {
 		t.Errorf("force should be false")
 	}
@@ -220,10 +197,10 @@ func TestConfigTestOK(t *testing.T) {
 	defer os.Remove(confFile.Name())
 
 	argv := []string{"-conf=" + confFile.Name()}
-	status := doConfigtest(argv)
+	err = doConfigtest(&flag.FlagSet{}, argv)
 
-	if status != exitStatusOK {
-		t.Errorf("configtest(ok) must be return exitStatusOK")
+	if err != nil {
+		t.Errorf("configtest(ok) must be return nil")
 	}
 }
 
@@ -240,10 +217,10 @@ func TestConfigTestNotFound(t *testing.T) {
 	defer os.Remove(confFile.Name())
 
 	argv := []string{"-conf=" + confFile.Name() + "xxx"}
-	status := doConfigtest(argv)
+	err = doConfigtest(&flag.FlagSet{}, argv)
 
-	if status != exitStatusError {
-		t.Errorf("configtest(failed) must be return existStatusError")
+	if err == nil {
+		t.Errorf("configtest(failed) must be return error")
 	}
 }
 
@@ -262,9 +239,23 @@ command = "bar"
 	defer os.Remove(confFile.Name())
 
 	argv := []string{"-conf=" + confFile.Name()}
-	status := doConfigtest(argv)
+	err = doConfigtest(&flag.FlagSet{}, argv)
 
-	if status != exitStatusError {
-		t.Errorf("configtest(failed) must be return exitStatusError")
+	if err == nil {
+		t.Errorf("configtest(failed) must be return error")
+	}
+}
+
+func TestDoOnce(t *testing.T) {
+	err := doOnce(&flag.FlagSet{}, []string{})
+	if err != nil {
+		t.Errorf("doOnce should return nil even if argv is empty, but returns %s", err)
+	}
+}
+
+func TestDoVersion(t *testing.T) {
+	err := doVersion(&flag.FlagSet{}, []string{})
+	if err != nil {
+		t.Errorf("doVersion should return nil, but returns %s", err)
 	}
 }