+++ util/util.go
@@ -13,18 +13,18 @@ import (
 
 var utilLogger = logging.GetLogger("util")
 
-// timeoutDuration is option of `Runcommand()` set timeout limit of command execution.
-var timeoutDuration = 30 * time.Second
+// TimeoutDuration is option of `Runcommand()` set timeout limit of command execution.
+var TimeoutDuration = 30 * time.Second
 
-// timeoutKillAfter is option of `RunCommand()` set waiting limit to `kill -kill` after terminating the command.
-var timeoutKillAfter = 10 * time.Second
+// TimeoutKillAfter is option of `RunCommand()` set waiting limit to `kill -kill` after terminating the command.
+var TimeoutKillAfter = 10 * time.Second
 
 // RunCommand runs command (in one string) and returns stdout, stderr strings and its exit code.
 func RunCommand(command string) (string, string, int, error) {
 	tio := &timeout.Timeout{
 		Cmd:       exec.Command("/bin/sh", "-c", command),
-		Duration:  timeoutDuration,
-		KillAfter: timeoutKillAfter,
+		Duration:  TimeoutDuration,
+		KillAfter: TimeoutKillAfter,
 	}
 	exitStatus, stdout, stderr, err := tio.Run()